+++ sdk/resourcemanager/compute/armcompute/models.go
@@ -368,7 +368,7 @@ type CapacityReservationGroupInstanceView struct {
 	// READ-ONLY; List of instance view of the capacity reservations under the capacity reservation group.
 	CapacityReservations []*CapacityReservationInstanceViewWithName
 
-	// READ-ONLY; List of the subscriptions that the capacity reservation group is shared with. Note: Minimum api-version: 2024-03-01.
+	// READ-ONLY; List of the subscriptions that the capacity reservation group is shared with. Note: Minimum api-version: 2023-09-01.
 	// Please refer to https://aka.ms/computereservationsharing for more details.
 	SharedSubscriptionIDs []*SubResourceReadOnly
 }
@@ -388,7 +388,7 @@ type CapacityReservationGroupProperties struct {
 	// Specifies the settings to enable sharing across subscriptions for the capacity reservation group resource. Pls. keep in
 	// mind the capacity reservation group resource generally can be shared across
 	// subscriptions belonging to a single azure AAD tenant or cross AAD tenant if there is a trust relationship established between
-	// the AAD tenants. Note: Minimum api-version: 2024-03-01. Please refer to
+	// the AAD tenants. Note: Minimum api-version: 2023-09-01. Please refer to
 	// https://aka.ms/computereservationsharing for more details.
 	SharingProfile *ResourceSharingProfile
 
@@ -4101,7 +4101,7 @@ type ProximityPlacementGroupUpdate struct {
 	Tags map[string]*string
 }
 
-// ProxyAgentSettings - Specifies ProxyAgent settings while creating the virtual machine. Minimum api-version: 2024-03-01.
+// ProxyAgentSettings - Specifies ProxyAgent settings while creating the virtual machine. Minimum api-version: 2023-09-01.
 type ProxyAgentSettings struct {
 	// Specifies whether ProxyAgent feature should be enabled on the virtual machine or virtual machine scale set.
 	Enabled *bool
@@ -4458,7 +4458,7 @@ type ResourceSKUsResult struct {
 
 type ResourceSharingProfile struct {
 	// Specifies an array of subscription resource IDs that capacity reservation group is shared with. Note: Minimum api-version:
-	// 2024-03-01. Please refer to https://aka.ms/computereservationsharing for more
+	// 2023-09-01. Please refer to https://aka.ms/computereservationsharing for more
 	// details.
 	SubscriptionIDs []*SubResource
 }
@@ -5195,7 +5195,7 @@ type SecurityProfile struct {
 	// Specifies the Managed Identity used by ADE to get access token for keyvault operations.
 	EncryptionIdentity *EncryptionIdentity
 
-	// Specifies ProxyAgent settings while creating the virtual machine. Minimum api-version: 2024-03-01.
+	// Specifies ProxyAgent settings while creating the virtual machine. Minimum api-version: 2023-09-01.
 	ProxyAgentSettings *ProxyAgentSettings
 
 	// Specifies the SecurityType of the virtual machine. It has to be set to any specified value to enable UefiSettings. The
@@ -7674,6 +7674,9 @@ type VirtualMachineScaleSetProperties struct {
 	// The virtual machine profile.
 	VirtualMachineProfile *VirtualMachineScaleSetVMProfile
 
+	// Specifies the align mode between Virtual Machine Scale Set compute and storage Fault Domain count.
+	ZonalPlatformFaultDomainAlignMode *ZonalPlatformFaultDomainAlignMode
+
 	// Whether to force strictly even Virtual Machine distribution cross x-zones in case there is zone outage. zoneBalance property
 	// can only be set if the zones property of the scale set contains more than
 	// one zone. If there are no zones or only one zone specified, then zoneBalance property should not be set.
@@ -8352,7 +8355,7 @@ type VirtualMachineScaleSetVMProfile struct {
 
 	// READ-ONLY; Specifies the time in which this VM profile for the Virtual Machine Scale Set was created. Minimum API version
 	// for this property is 2024-03-01. This value will be added to VMSS Flex VM tags when
-	// creating/updating the VMSS VM Profile with minimum api-version 2024-03-01.
+	// creating/updating the VMSS VM Profile with minimum api-version 2023-09-01.
 	TimeCreated *time.Time
 }